+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased][]
 
+## [1.1.0][] - 2021-01-14
+
 ### Changed
 
 -   Use `enhanced-resolve` resolver for `postcss-import`
@@ -17,3 +19,6 @@
 -   Initial implementation
 
 [1.0.0]: https://github.com/niksy/sass-module-resolve-importer/tree/v1.0.0
+[unreleased]:
+	https://github.com/niksy/sass-module-resolve-importer/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/niksy/sass-module-resolve-importer/tree/v1.1.0